requirement.txt
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.4.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
marshmallow==3.7.0
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
passlib==1.7.2
pep517==0.8.2
progress==1.5
pydotenv==0.0.7
pylint==2.5.3
pyparsing==2.4.6
python-dotenv==0.14.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
